fix(#906): address remaining gaps — realtime events, back-compat shim, file renames, API v1 boundary
- GAP 1: Rename realtime event strings generation.frame:created → generation.shot:created,
  generation.frame:updated → generation.shot:updated in both emitters (scene-split-workflow.ts,
  motion-prompt-scene-workflow.ts, visual-prompt-scene-workflow.ts) and consumers
  (query-cache-updater.ts, use-generation-stream.ts). Also rename 'frame:created'/'frame:updated'
  keys in realtimeSchema (index.ts) to 'shot:created'/'shot:updated'.

- GAP 2: Add back-compat shim in runImpl of image-workflow, shot-variant-workflow,
  upscale-shot-variant-workflow, and motion-workflow — accepts frameId from in-flight CF Workflow
  instances serialized before #906. Falls back via: shotId ?? frameId ?? undefined.

- GAP 4: git mv frame-staleness-banners.tsx → shot-staleness-banners.tsx; update import in
  scene-script-prompts.tsx and comment in sheet-staleness-banners.tsx.

- GAP 5: Rename generateFrameMotionFn → generateShotMotionFn in motion-functions.ts and
  its callers (scene-script-prompts.tsx, rescan-continuity-from-prompt.ts).
  Keep frame.schemas.ts, build-frame-image-input.ts, use-frame-prompt-stream.ts, and
  api/test/frame.ts as intentional (literal still-image / test-seed compatibility).

- GAP 7: "Editing frames..." → "Editing shots..." and "Used in N frame(s)" → "Used in N shot(s)"
  in element-card.tsx.

- GAP 3: Confirmed correct — api-v1/state.ts wire field frames intentionally kept (external
  contract); internal type SequenceStateShot; DB accessor scopedDb.shots.
- GAP 6: Confirmed correct — staleness-indicator.tsx entityType union only has 'shot'.
