+++ src/components/scenes/shot-staleness-banners.tsx
